add production and development config files and load them from start file.
add two config files for development and production which contains
associative arrays of key-value pairs. mode.php contains the
config to be used for the current environment. It is currently
development. This mode is loaded in the start file and then the
config can be used by the config package.