add Password Hashing helper and use a singleton to use the Helper class
A hash class is created inside a Helpers dir where password_hash
and password_verify functions are used. The algorithm and cost
are pulled from the config defined previously.

A namespace is used to import the Hash helper file in the app and
a singleton 'hash' is created that infers the Hash class.

Additionally, a flash route is created which shows a message and
then redirects back to home. This will be helpful to show messages
to users.